Crate markovExperimental

A generic Markov chain for almost any type. This uses HashMaps internally, and so Eq and Hash are both required.

Examples

use markov::Chain;
 
let mut chain = Chain::for_strings();
chain.feed_str("I like cats and I like dogs.");
println!("{}", chain.generate_str());

use markov::Chain;

let mut chain = Chain::new(0u8, 255);
chain.feed(vec![1u8, 2, 3, 5]).feed(vec![3u8, 9, 2]);
println!("{:?}", chain.generate());

Structs

Chain
A generic Markov chain for almost any type. This uses HashMaps internally, and so Eq and Hash are both required.

InfiniteChainIterator
An infinite iterator over a Markov chain.

SizedChainIterator
A sized iterator over a Markov chain.

Traits

Chainable
The definition of all types that can be used in a Chain.

Type Definitions

InfiniteChainStringIterator
An infinite iterator over a Markov chain of strings.

SizedChainStringIterator
A sized iterator over a Markov chain of strings.